refactor: peel four submodules off aggregate_state.rs
aggregate_state.rs was a 3991-line god-file doing six separable jobs.
This lifts four of them into submodules; the remaining monolith is
the IncrementalAggState implementation itself (~1800 lines), which
is a bigger refactor on its own.

New submodules under src/aggregate_state/:
- scalar_json.rs  (153 lines)  ScalarValue ↔ serde_json::Value for
                                checkpoint serialisation.
- checkpoints.rs   (78 lines)  On-disk checkpoint shapes for group,
                                window, and join state + the
                                query_fingerprint helper.
- keys.rs          (74 lines)  Group-key derivation between Arrow's
                                OwnedRow and Vec<ScalarValue>.
- compile.rs      (539 lines)  LogicalPlan introspection,
                                expr-to-SQL, HAVING compilation,
                                SqlClauses extraction,
                                CompiledProjection.

Main file: 3991 → 3212 lines (−779, ~20% smaller). No behaviour
change; all 2253 lib tests pass, clippy + fmt clean.

AggregateInfo and SqlClauses were flagged as unused outside compile.rs
by the compiler and dropped from the crate re-export surface.
