+++ crates/laminar-db/src/aggregate_state.rs
@@ -1,6 +1,5 @@
 //! Incremental aggregation state for streaming GROUP BY queries.
 
-use std::hash::{Hash, Hasher};
 use std::sync::Arc;
 
 use ahash::AHashMap;
@@ -12,77 +11,26 @@ use arrow::datatypes::{DataType, Field, Schema, SchemaRef};
 use arrow::record_batch::RecordBatch;
 use datafusion::physical_expr::{create_physical_expr, PhysicalExpr};
 use datafusion::prelude::SessionContext;
-use datafusion_common::{DFSchema, ScalarValue};
+use datafusion_common::ScalarValue;
 use datafusion_expr::function::AccumulatorArgs;
 use datafusion_expr::{AggregateUDF, LogicalPlan};
-use serde_json::json;
 
 use crate::error::DbError;
 
-pub(crate) fn row_to_scalar_key_with_types(
-    converter: &arrow::row::RowConverter,
-    row_key: &arrow::row::OwnedRow,
-    group_types: &[DataType],
-) -> Result<Vec<ScalarValue>, DbError> {
-    let row_as_cols = converter
-        .convert_rows(std::iter::once(row_key.row()))
-        .map_err(|e| DbError::Pipeline(format!("row→key: {e}")))?;
-
-    let mut sv_key = Vec::with_capacity(group_types.len());
-    for (col_idx, arr) in row_as_cols.iter().enumerate() {
-        let sv = ScalarValue::try_from_array(arr, 0)
-            .map_err(|e| DbError::Pipeline(format!("group key decode: {e}")))?;
-        if sv.data_type() == group_types[col_idx] {
-            sv_key.push(sv);
-        } else {
-            sv_key.push(sv.cast_to(&group_types[col_idx]).unwrap_or(sv));
-        }
-    }
-    Ok(sv_key)
-}
-
-pub(crate) fn global_aggregate_key() -> arrow::row::OwnedRow {
-    use std::sync::OnceLock;
-    static SENTINEL: OnceLock<arrow::row::OwnedRow> = OnceLock::new();
-    SENTINEL
-        .get_or_init(|| {
-            let conv =
-                arrow::row::RowConverter::new(vec![arrow::row::SortField::new(DataType::Boolean)])
-                    .unwrap();
-            let col: ArrayRef = Arc::new(arrow::array::BooleanArray::from(vec![false]));
-            conv.convert_columns(&[col]).unwrap().row(0).owned()
-        })
-        .clone()
-}
-
-pub(crate) fn scalar_key_to_owned_row(
-    row_converter: &arrow::row::RowConverter,
-    sv_key: &[ScalarValue],
-    target_types: &[DataType],
-) -> Result<arrow::row::OwnedRow, DbError> {
-    if sv_key.is_empty() {
-        return Ok(global_aggregate_key());
-    }
-    let arrays: Vec<ArrayRef> = sv_key
-        .iter()
-        .enumerate()
-        .map(|(i, sv)| {
-            let arr = sv
-                .to_array()
-                .map_err(|e| DbError::Pipeline(format!("scalar to array: {e}")))?;
-            if i < target_types.len() && arr.data_type() != &target_types[i] {
-                arrow::compute::cast(&arr, &target_types[i])
-                    .map_err(|e| DbError::Pipeline(format!("key type cast: {e}")))
-            } else {
-                Ok(arr)
-            }
-        })
-        .collect::<Result<_, _>>()?;
-    let rows = row_converter
-        .convert_columns(&arrays)
-        .map_err(|e| DbError::Pipeline(format!("key to row: {e}")))?;
-    Ok(rows.row(0).owned())
-}
+mod checkpoints;
+mod compile;
+mod keys;
+mod scalar_json;
+pub(crate) use checkpoints::{
+    query_fingerprint, AggStateCheckpoint, EmittedCheckpoint, EowcStateCheckpoint,
+    GroupCheckpoint, JoinStateCheckpoint, WindowCheckpoint,
+};
+pub(crate) use compile::{
+    apply_compiled_having, compile_having_filter, expr_to_sql, extract_clauses, find_aggregate,
+    resolve_expr_type, CompiledProjection,
+};
+pub(crate) use keys::{global_aggregate_key, row_to_scalar_key_with_types, scalar_key_to_owned_row};
+pub(crate) use scalar_json::{json_to_scalar, scalar_to_json};
 
 pub(crate) fn emit_window_batch(
     window_start: i64,
@@ -150,216 +98,6 @@ pub(crate) fn emit_window_batch(
     Ok(Some(batch))
 }
 
-pub(crate) fn scalar_to_json(sv: &ScalarValue) -> serde_json::Value {
-    match sv {
-        ScalarValue::Null => json!({"t": "N"}),
-        ScalarValue::Boolean(None) => json!({"t": "B", "v": null}),
-        ScalarValue::Boolean(Some(b)) => json!({"t": "B", "v": b}),
-        ScalarValue::Int8(None)
-        | ScalarValue::Int16(None)
-        | ScalarValue::Int32(None)
-        | ScalarValue::Int64(None) => json!({"t": "I64", "v": null}),
-        ScalarValue::Int8(Some(n)) => json!({"t": "I64", "v": i64::from(*n)}),
-        ScalarValue::Int16(Some(n)) => json!({"t": "I64", "v": i64::from(*n)}),
-        ScalarValue::Int32(Some(n)) => json!({"t": "I64", "v": i64::from(*n)}),
-        ScalarValue::Int64(Some(n)) => json!({"t": "I64", "v": n}),
-        ScalarValue::UInt8(None)
-        | ScalarValue::UInt16(None)
-        | ScalarValue::UInt32(None)
-        | ScalarValue::UInt64(None) => json!({"t": "U64", "v": null}),
-        ScalarValue::UInt8(Some(n)) => json!({"t": "U64", "v": u64::from(*n)}),
-        ScalarValue::UInt16(Some(n)) => json!({"t": "U64", "v": u64::from(*n)}),
-        ScalarValue::UInt32(Some(n)) => json!({"t": "U64", "v": u64::from(*n)}),
-        ScalarValue::UInt64(Some(n)) => json!({"t": "U64", "v": n}),
-        ScalarValue::Float32(None) | ScalarValue::Float64(None) => {
-            json!({"t": "F64", "v": null})
-        }
-        ScalarValue::Float32(Some(f)) => json!({"t": "F64", "v": f64::from(*f)}),
-        ScalarValue::Float64(Some(f)) => json!({"t": "F64", "v": f}),
-        ScalarValue::Utf8(None) | ScalarValue::LargeUtf8(None) | ScalarValue::Utf8View(None) => {
-            json!({"t": "S", "v": null})
-        }
-        ScalarValue::Utf8(Some(s))
-        | ScalarValue::LargeUtf8(Some(s))
-        | ScalarValue::Utf8View(Some(s)) => json!({"t": "S", "v": s}),
-        ScalarValue::TimestampNanosecond(v, tz) => {
-            json!({"t": "TSns", "v": v, "tz": tz.as_ref().map(ToString::to_string)})
-        }
-        ScalarValue::TimestampMicrosecond(v, tz) => {
-            json!({"t": "TSus", "v": v, "tz": tz.as_ref().map(ToString::to_string)})
-        }
-        ScalarValue::TimestampMillisecond(v, tz) => {
-            json!({"t": "TSms", "v": v, "tz": tz.as_ref().map(ToString::to_string)})
-        }
-        ScalarValue::TimestampSecond(v, tz) => {
-            json!({"t": "TSs", "v": v, "tz": tz.as_ref().map(ToString::to_string)})
-        }
-        ScalarValue::Date32(v) => json!({"t": "D32", "v": v}),
-        ScalarValue::Date64(v) => json!({"t": "D64", "v": v}),
-        ScalarValue::List(arr) => {
-            use arrow::array::Array;
-            let list_arr: Option<&arrow::array::ListArray> = arr.as_any().downcast_ref();
-            match list_arr {
-                Some(list) if !list.is_empty() => {
-                    let values = list.value(0);
-                    let mut items = Vec::with_capacity(values.len());
-                    for i in 0..values.len() {
-                        let sv =
-                            ScalarValue::try_from_array(&values, i).unwrap_or(ScalarValue::Null);
-                        items.push(scalar_to_json(&sv));
-                    }
-                    json!({"t": "L", "v": items})
-                }
-                _ => json!({"t": "L", "v": []}),
-            }
-        }
-        // Fallback: encode as string representation with type tag
-        other => json!({"t": "STR", "v": other.to_string()}),
-    }
-}
-
-pub(crate) fn json_to_scalar(v: &serde_json::Value) -> Result<ScalarValue, DbError> {
-    let t = v
-        .get("t")
-        .and_then(|t| t.as_str())
-        .ok_or_else(|| DbError::Pipeline("missing type tag in scalar JSON".to_string()))?;
-    let val = v.get("v");
-    match t {
-        "N" => Ok(ScalarValue::Null),
-        "B" => match val.and_then(serde_json::Value::as_bool) {
-            Some(b) => Ok(ScalarValue::Boolean(Some(b))),
-            None => Ok(ScalarValue::Boolean(None)),
-        },
-        "I64" => match val {
-            Some(serde_json::Value::Number(n)) => Ok(ScalarValue::Int64(n.as_i64())),
-            _ => Ok(ScalarValue::Int64(None)),
-        },
-        "U64" => match val {
-            Some(serde_json::Value::Number(n)) => Ok(ScalarValue::UInt64(n.as_u64())),
-            _ => Ok(ScalarValue::UInt64(None)),
-        },
-        "F64" => match val {
-            Some(serde_json::Value::Number(n)) => Ok(ScalarValue::Float64(n.as_f64())),
-            _ => Ok(ScalarValue::Float64(None)),
-        },
-        "S" => match val.and_then(|v| v.as_str()) {
-            Some(s) => Ok(ScalarValue::Utf8(Some(s.to_string()))),
-            None => Ok(ScalarValue::Utf8(None)),
-        },
-        "TSns" | "TSus" | "TSms" | "TSs" => {
-            let ts = val.and_then(serde_json::Value::as_i64);
-            let tz: Option<Arc<str>> = v.get("tz").and_then(|v| v.as_str()).map(Arc::from);
-            match t {
-                "TSns" => Ok(ScalarValue::TimestampNanosecond(ts, tz)),
-                "TSus" => Ok(ScalarValue::TimestampMicrosecond(ts, tz)),
-                "TSms" => Ok(ScalarValue::TimestampMillisecond(ts, tz)),
-                _ => Ok(ScalarValue::TimestampSecond(ts, tz)),
-            }
-        }
-        "D32" => {
-            #[allow(clippy::cast_possible_truncation)]
-            let d = val.and_then(serde_json::Value::as_i64).map(|n| n as i32);
-            Ok(ScalarValue::Date32(d))
-        }
-        "D64" => {
-            let d = val.and_then(serde_json::Value::as_i64);
-            Ok(ScalarValue::Date64(d))
-        }
-        "L" => {
-            let items = val
-                .and_then(|v| v.as_array())
-                .ok_or_else(|| DbError::Pipeline("expected array for List scalar".to_string()))?;
-            let scalars: Result<Vec<ScalarValue>, _> = items.iter().map(json_to_scalar).collect();
-            let scalars = scalars?;
-            if scalars.is_empty() {
-                Ok(ScalarValue::List(Arc::new(
-                    arrow::array::GenericListArray::new_null(
-                        Arc::new(Field::new("item", DataType::Null, true)),
-                        1,
-                    ),
-                )))
-            } else {
-                let arr = ScalarValue::new_list(&scalars, &scalars[0].data_type(), true);
-                Ok(ScalarValue::List(arr))
-            }
-        }
-        "STR" => match val.and_then(|v| v.as_str()) {
-            Some(s) => Ok(ScalarValue::Utf8(Some(s.to_string()))),
-            None => Ok(ScalarValue::Utf8(None)),
-        },
-        other => Err(DbError::Pipeline(format!(
-            "unsupported scalar type tag in checkpoint: {other}"
-        ))),
-    }
-}
-
-pub(crate) fn query_fingerprint(pre_agg_sql: &str, output_schema: &Schema) -> u64 {
-    let mut hasher = std::hash::DefaultHasher::new();
-    pre_agg_sql.hash(&mut hasher);
-    for field in output_schema.fields() {
-        field.name().hash(&mut hasher);
-        field.data_type().to_string().hash(&mut hasher);
-    }
-    hasher.finish()
-}
-
-#[derive(Clone, serde::Serialize, serde::Deserialize)]
-pub(crate) struct GroupCheckpoint {
-    pub key: Vec<serde_json::Value>,
-    pub acc_states: Vec<Vec<serde_json::Value>>,
-    #[serde(default = "default_last_updated")]
-    pub last_updated_ms: i64,
-}
-
-fn default_last_updated() -> i64 {
-    i64::MIN
-}
-
-#[derive(Clone, serde::Serialize, serde::Deserialize)]
-pub(crate) struct AggStateCheckpoint {
-    pub fingerprint: u64,
-    pub groups: Vec<GroupCheckpoint>,
-    #[serde(default)]
-    pub last_emitted: Vec<EmittedCheckpoint>,
-}
-
-#[derive(Clone, serde::Serialize, serde::Deserialize)]
-pub(crate) struct EmittedCheckpoint {
-    pub key: Vec<serde_json::Value>,
-    pub values: Vec<serde_json::Value>,
-}
-
-#[derive(Clone, serde::Serialize, serde::Deserialize)]
-pub(crate) struct WindowCheckpoint {
-    pub window_start: i64,
-    pub groups: Vec<GroupCheckpoint>,
-}
-
-#[derive(Clone, serde::Serialize, serde::Deserialize)]
-pub(crate) struct EowcStateCheckpoint {
-    pub fingerprint: u64,
-    pub windows: Vec<WindowCheckpoint>,
-}
-
-#[derive(Clone, serde::Serialize, serde::Deserialize)]
-pub(crate) struct JoinStateCheckpoint {
-    #[serde(default)]
-    pub left_buffer_rows: u64,
-    #[serde(default)]
-    pub right_buffer_rows: u64,
-    #[serde(default)]
-    pub left_batches: Vec<Vec<u8>>,
-    #[serde(default)]
-    pub right_batches: Vec<Vec<u8>>,
-    #[serde(default = "default_evicted_watermark")]
-    pub last_evicted_watermark: i64,
-    #[serde(default = "default_evicted_watermark")]
-    pub last_evicted_watermark_right: i64,
-}
-
-fn default_evicted_watermark() -> i64 {
-    i64::MIN
-}
 
 pub(crate) struct AggFuncSpec {
     pub(crate) udf: Arc<AggregateUDF>,
@@ -1441,526 +1179,9 @@ impl IncrementalAggState {
     }
 }
 
-/// Result of finding an aggregate node in a logical plan.
-pub(crate) struct AggregateInfo {
-    pub(crate) group_exprs: Vec<datafusion_expr::Expr>,
-    pub(crate) aggr_exprs: Vec<datafusion_expr::Expr>,
-    pub(crate) schema: Arc<Schema>,
-    /// Input schema (pre-widening) for resolving aggregate argument types.
-    pub(crate) input_schema: Arc<Schema>,
-    /// HAVING predicate (from a Filter node directly above the Aggregate).
-    pub(crate) having_predicate: Option<datafusion_expr::Expr>,
-    /// `DFSchema` of the Aggregate's input (for compiling pre-agg expressions).
-    pub(crate) input_df_schema: Arc<DFSchema>,
-    /// WHERE predicate from a Filter node below the Aggregate, if any.
-    pub(crate) where_predicate: Option<datafusion_expr::Expr>,
-}
-
-pub(crate) fn find_aggregate(plan: &LogicalPlan) -> Option<AggregateInfo> {
-    find_aggregate_inner(plan, None)
-}
-
-fn find_aggregate_inner(
-    plan: &LogicalPlan,
-    parent_filter: Option<&datafusion_expr::Expr>,
-) -> Option<AggregateInfo> {
-    match plan {
-        LogicalPlan::Aggregate(agg) => {
-            let schema = Arc::new(agg.schema.as_arrow().clone());
-            let input_schema = Arc::new(agg.input.schema().as_arrow().clone());
-            let input_df_schema = Arc::clone(agg.input.schema());
-            let where_predicate = extract_where_predicate(&agg.input);
-            Some(AggregateInfo {
-                group_exprs: agg.group_expr.clone(),
-                aggr_exprs: agg.aggr_expr.clone(),
-                schema,
-                input_schema,
-                having_predicate: parent_filter.cloned(),
-                input_df_schema,
-                where_predicate,
-            })
-        }
-        // A Filter directly above an Aggregate is a HAVING clause
-        LogicalPlan::Filter(filter) => {
-            if matches!(&*filter.input, LogicalPlan::Aggregate(_)) {
-                find_aggregate_inner(&filter.input, Some(&filter.predicate))
-            } else {
-                find_aggregate_inner(&filter.input, None)
-            }
-        }
-        // Walk through wrappers that don't change aggregation semantics
-        LogicalPlan::Projection(proj) => find_aggregate_inner(&proj.input, None),
-        LogicalPlan::Sort(sort) => find_aggregate_inner(&sort.input, None),
-        LogicalPlan::Limit(limit) => find_aggregate_inner(&limit.input, None),
-        LogicalPlan::SubqueryAlias(alias) => find_aggregate_inner(&alias.input, None),
-        _ => {
-            for input in plan.inputs() {
-                if let Some(result) = find_aggregate_inner(input, None) {
-                    return Some(result);
-                }
-            }
-            None
-        }
-    }
-}
-
-fn extract_where_predicate(plan: &LogicalPlan) -> Option<datafusion_expr::Expr> {
-    match plan {
-        LogicalPlan::Filter(f) => Some(f.predicate.clone()),
-        LogicalPlan::Projection(p) => extract_where_predicate(&p.input),
-        LogicalPlan::Sort(s) => extract_where_predicate(&s.input),
-        LogicalPlan::Limit(l) => extract_where_predicate(&l.input),
-        LogicalPlan::SubqueryAlias(a) => extract_where_predicate(&a.input),
-        _ => None,
-    }
-}
-
-fn scalar_value_to_sql(sv: &ScalarValue) -> String {
-    match sv {
-        ScalarValue::Utf8(Some(s))
-        | ScalarValue::LargeUtf8(Some(s))
-        | ScalarValue::Utf8View(Some(s)) => {
-            format!("'{}'", s.replace('\'', "''"))
-        }
-        ScalarValue::Utf8(None)
-        | ScalarValue::LargeUtf8(None)
-        | ScalarValue::Utf8View(None)
-        | ScalarValue::Null
-        | ScalarValue::Boolean(None) => "NULL".to_string(),
-        ScalarValue::Boolean(Some(b)) => if *b { "TRUE" } else { "FALSE" }.to_string(),
-        ScalarValue::IntervalDayTime(Some(v)) => {
-            let mut parts = Vec::new();
-            if v.days != 0 {
-                parts.push(format!("{} days", v.days));
-            }
-            if v.milliseconds != 0 || parts.is_empty() {
-                let abs_ms = v.milliseconds.unsigned_abs();
-                let secs = abs_ms / 1000;
-                let frac = abs_ms % 1000;
-                let sign = if v.milliseconds < 0 { "-" } else { "" };
-                if frac == 0 {
-                    parts.push(format!("{sign}{secs} seconds"));
-                } else {
-                    parts.push(format!("{sign}{secs}.{frac:03} seconds"));
-                }
-            }
-            format!("INTERVAL '{}'", parts.join(" "))
-        }
-        ScalarValue::IntervalYearMonth(Some(v)) => {
-            let years = v / 12;
-            let months = v % 12;
-            let mut parts = Vec::new();
-            if years != 0 {
-                parts.push(format!("{years} years"));
-            }
-            if months != 0 || parts.is_empty() {
-                parts.push(format!("{months} months"));
-            }
-            format!("INTERVAL '{}'", parts.join(" "))
-        }
-        ScalarValue::IntervalMonthDayNano(Some(v)) => {
-            let mut parts = Vec::new();
-            if v.months != 0 {
-                parts.push(format!("{} months", v.months));
-            }
-            if v.days != 0 {
-                parts.push(format!("{} days", v.days));
-            }
-            let nanos = v.nanoseconds;
-            if nanos != 0 || parts.is_empty() {
-                let abs_ns = nanos.unsigned_abs();
-                let secs = abs_ns / 1_000_000_000;
-                let remainder_ns = abs_ns % 1_000_000_000;
-                let sign = if nanos < 0 { "-" } else { "" };
-                if remainder_ns == 0 {
-                    parts.push(format!("{sign}{secs} seconds"));
-                } else {
-                    let millis = remainder_ns / 1_000_000;
-                    parts.push(format!("{sign}{secs}.{millis:03} seconds"));
-                }
-            }
-            format!("INTERVAL '{}'", parts.join(" "))
-        }
-        _ => sv.to_string(),
-    }
-}
-
-fn case_to_sql(case: &datafusion_expr::expr::Case) -> String {
-    use std::fmt::Write;
-    let mut sql = String::from("CASE");
-    if let Some(operand) = &case.expr {
-        let _ = write!(sql, " {}", expr_to_sql(operand));
-    }
-    for (when_expr, then_expr) in &case.when_then_expr {
-        let _ = write!(
-            sql,
-            " WHEN {} THEN {}",
-            expr_to_sql(when_expr),
-            expr_to_sql(then_expr)
-        );
-    }
-    if let Some(else_expr) = &case.else_expr {
-        let _ = write!(sql, " ELSE {}", expr_to_sql(else_expr));
-    }
-    sql.push_str(" END");
-    sql
-}
-
-pub(crate) fn expr_to_sql(expr: &datafusion_expr::Expr) -> String {
-    use datafusion_expr::Expr;
-    match expr {
-        Expr::Column(col) => format!("\"{}\"", col.name),
-        Expr::Literal(sv, _) => scalar_value_to_sql(sv),
-        Expr::Alias(alias) => expr_to_sql(&alias.expr),
-        Expr::BinaryExpr(bin) => {
-            let left = expr_to_sql(&bin.left);
-            let right = expr_to_sql(&bin.right);
-            format!("({left} {op} {right})", op = bin.op)
-        }
-        Expr::Cast(cast) => {
-            let inner = expr_to_sql(&cast.expr);
-            format!("CAST({inner} AS {})", cast.data_type)
-        }
-        Expr::TryCast(cast) => {
-            let inner = expr_to_sql(&cast.expr);
-            format!("TRY_CAST({inner} AS {})", cast.data_type)
-        }
-        Expr::ScalarFunction(func) => {
-            let args: Vec<String> = func.args.iter().map(expr_to_sql).collect();
-            format!("{}({})", func.func.name(), args.join(", "))
-        }
-        Expr::AggregateFunction(agg) => {
-            let name = agg.func.name();
-            let args: Vec<String> = agg.params.args.iter().map(expr_to_sql).collect();
-            if agg.params.distinct {
-                format!("{name}(DISTINCT {})", args.join(", "))
-            } else {
-                format!("{name}({})", args.join(", "))
-            }
-        }
-        Expr::Case(case) => case_to_sql(case),
-        Expr::Not(inner) => format!("(NOT {})", expr_to_sql(inner)),
-        Expr::Negative(inner) => format!("(-{})", expr_to_sql(inner)),
-        Expr::IsNull(inner) => {
-            format!("({} IS NULL)", expr_to_sql(inner))
-        }
-        Expr::IsNotNull(inner) => {
-            format!("({} IS NOT NULL)", expr_to_sql(inner))
-        }
-        Expr::IsTrue(inner) => {
-            format!("({} IS TRUE)", expr_to_sql(inner))
-        }
-        Expr::IsFalse(inner) => {
-            format!("({} IS FALSE)", expr_to_sql(inner))
-        }
-        Expr::IsNotTrue(inner) => {
-            format!("({} IS NOT TRUE)", expr_to_sql(inner))
-        }
-        Expr::IsNotFalse(inner) => {
-            format!("({} IS NOT FALSE)", expr_to_sql(inner))
-        }
-        Expr::Between(between) => {
-            let e = expr_to_sql(&between.expr);
-            let low = expr_to_sql(&between.low);
-            let high = expr_to_sql(&between.high);
-            let not = if between.negated { " NOT" } else { "" };
-            format!("({e}{not} BETWEEN {low} AND {high})")
-        }
-        Expr::InList(in_list) => {
-            let e = expr_to_sql(&in_list.expr);
-            let items: Vec<String> = in_list.list.iter().map(expr_to_sql).collect();
-            let not = if in_list.negated { " NOT" } else { "" };
-            format!("({e}{not} IN ({}))", items.join(", "))
-        }
-        Expr::Like(like) => {
-            let e = expr_to_sql(&like.expr);
-            let pat = expr_to_sql(&like.pattern);
-            let kw = if like.case_insensitive {
-                "ILIKE"
-            } else {
-                "LIKE"
-            };
-            let not = if like.negated { " NOT" } else { "" };
-            if let Some(esc) = &like.escape_char {
-                format!("({e}{not} {kw} {pat} ESCAPE '{esc}')")
-            } else {
-                format!("({e}{not} {kw} {pat})")
-            }
-        }
-        #[allow(deprecated)]
-        Expr::Wildcard { .. } => "TRUE".to_string(),
-        other => other.to_string(),
-    }
-}
-
-/// Pre-compiled projection for evaluating pre-agg expressions without SQL.
-///
-/// Replaces the `ctx.sql(pre_agg_sql).await.collect().await` hot-path call
-/// with direct `PhysicalExpr::evaluate()` on the source batch, eliminating
-/// SQL parsing, logical planning, physical planning, and `MemTable` overhead.
-pub(crate) struct CompiledProjection {
-    #[allow(dead_code)]
-    pub(crate) source_table: String,
-    pub(crate) exprs: Vec<Arc<dyn PhysicalExpr>>,
-    pub(crate) filter: Option<Arc<dyn PhysicalExpr>>,
-    pub(crate) output_schema: SchemaRef,
-}
-
-impl CompiledProjection {
-    /// Evaluate the projection against a source batch.
-    ///
-    /// Applies the WHERE filter (if any), then evaluates each expression
-    /// to produce the projected output batch.
-    pub(crate) fn evaluate(&self, batch: &RecordBatch) -> Result<RecordBatch, DbError> {
-        if batch.num_rows() == 0 {
-            return Ok(RecordBatch::new_empty(Arc::clone(&self.output_schema)));
-        }
-
-        // Apply WHERE filter first
-        let filtered = if let Some(ref filter) = self.filter {
-            let result = filter
-                .evaluate(batch)
-                .map_err(|e| DbError::Pipeline(format!("WHERE filter evaluate: {e}")))?;
-            let mask = result
-                .into_array(batch.num_rows())
-                .map_err(|e| DbError::Pipeline(format!("WHERE filter to array: {e}")))?;
-            let bool_arr = mask
-                .as_any()
-                .downcast_ref::<arrow::array::BooleanArray>()
-                .ok_or_else(|| DbError::Pipeline("WHERE filter not boolean".into()))?;
-            arrow::compute::filter_record_batch(batch, bool_arr)
-                .map_err(|e| DbError::Pipeline(format!("WHERE filter: {e}")))?
-        } else {
-            batch.clone()
-        };
-
-        if filtered.num_rows() == 0 {
-            return Ok(RecordBatch::new_empty(Arc::clone(&self.output_schema)));
-        }
-
-        // Evaluate each projection expression
-        let mut arrays = Vec::with_capacity(self.exprs.len());
-        for expr in &self.exprs {
-            let result = expr
-                .evaluate(&filtered)
-                .map_err(|e| DbError::Pipeline(format!("projection evaluate: {e}")))?;
-            let arr = result
-                .into_array(filtered.num_rows())
-                .map_err(|e| DbError::Pipeline(format!("projection to array: {e}")))?;
-            arrays.push(arr);
-        }
-
-        RecordBatch::try_new(Arc::clone(&self.output_schema), arrays)
-            .map_err(|e| DbError::Pipeline(format!("projection batch build: {e}")))
-    }
-}
-
-pub(crate) fn apply_compiled_having(
-    batches: &[RecordBatch],
-    having_filter: &Arc<dyn PhysicalExpr>,
-) -> Result<Vec<RecordBatch>, DbError> {
-    let mut result = Vec::with_capacity(batches.len());
-    for batch in batches {
-        if batch.num_rows() == 0 {
-            continue;
-        }
-        let mask_result = having_filter
-            .evaluate(batch)
-            .map_err(|e| DbError::Pipeline(format!("HAVING evaluate: {e}")))?;
-        let mask = mask_result
-            .into_array(batch.num_rows())
-            .map_err(|e| DbError::Pipeline(format!("HAVING to array: {e}")))?;
-        let bool_arr = mask
-            .as_any()
-            .downcast_ref::<arrow::array::BooleanArray>()
-            .ok_or_else(|| DbError::Pipeline("HAVING filter not boolean".into()))?;
-        let filtered = arrow::compute::filter_record_batch(batch, bool_arr)
-            .map_err(|e| DbError::Pipeline(format!("HAVING filter: {e}")))?;
-        if filtered.num_rows() > 0 {
-            result.push(filtered);
-        }
-    }
-    Ok(result)
-}
-
-pub(crate) fn compile_having_filter(
-    ctx: &SessionContext,
-    having_predicate: Option<&datafusion_expr::Expr>,
-    output_schema: &SchemaRef,
-) -> Option<Arc<dyn PhysicalExpr>> {
-    let having_pred = having_predicate?;
-    let df_schema = DFSchema::try_from(output_schema.as_ref().clone()).ok()?;
-    let state = ctx.state();
-    let props = state.execution_props();
-    create_physical_expr(having_pred, &df_schema, props).ok()
-}
-
-pub(crate) struct SqlClauses {
-    pub(crate) from_clause: String,
-    pub(crate) where_clause: String,
-}
-
-pub(crate) fn extract_clauses(sql: &str) -> SqlClauses {
-    if let Ok(clauses) = extract_clauses_ast(sql) {
-        return clauses;
-    }
-    // Fallback: heuristic extraction for non-standard SQL
-    SqlClauses {
-        from_clause: extract_from_clause_heuristic(sql),
-        where_clause: extract_where_clause_heuristic(sql),
-    }
-}
-
-fn extract_clauses_ast(sql: &str) -> Result<SqlClauses, DbError> {
-    use sqlparser::dialect::GenericDialect;
-    use sqlparser::parser::Parser;
-
-    let dialect = GenericDialect {};
-    let stmts = Parser::parse_sql(&dialect, sql)
-        .map_err(|e| DbError::Pipeline(format!("SQL parse error: {e}")))?;
-
-    let stmt = stmts
-        .into_iter()
-        .next()
-        .ok_or_else(|| DbError::Pipeline("empty SQL statement".to_string()))?;
-
-    let sqlparser::ast::Statement::Query(query) = stmt else {
-        return Err(DbError::Pipeline("expected SELECT statement".to_string()));
-    };
-
-    let sqlparser::ast::SetExpr::Select(select) = *query.body else {
-        return Err(DbError::Pipeline("expected simple SELECT".to_string()));
-    };
-
-    let from_clause = select
-        .from
-        .iter()
-        .map(ToString::to_string)
-        .collect::<Vec<_>>()
-        .join(", ");
-
-    let where_clause = select
-        .selection
-        .as_ref()
-        .map(|expr| format!(" WHERE {expr}"))
-        .unwrap_or_default();
-
-    Ok(SqlClauses {
-        from_clause,
-        where_clause,
-    })
-}
-
-fn extract_from_clause_heuristic(sql: &str) -> String {
-    let upper = sql.to_uppercase();
-    let from_pos = upper.find(" FROM ").map(|p| p + 6);
-    let Some(start) = from_pos else {
-        return String::new();
-    };
-    let rest = &sql[start..];
-    let end_keywords = [" WHERE ", " GROUP ", " ORDER ", " LIMIT ", " HAVING "];
-    let end = end_keywords
-        .iter()
-        .filter_map(|kw| rest.to_uppercase().find(kw))
-        .min()
-        .unwrap_or(rest.len());
-    rest[..end].trim().to_string()
-}
-
-fn extract_where_clause_heuristic(sql: &str) -> String {
-    let upper = sql.to_uppercase();
-    let where_pos = upper.find(" WHERE ");
-    let Some(start) = where_pos else {
-        return String::new();
-    };
-    let rest = &sql[start..];
-    let end_keywords = [" GROUP ", " ORDER ", " LIMIT ", " HAVING "];
-    let end = end_keywords
-        .iter()
-        .filter_map(|kw| rest[7..].to_uppercase().find(kw).map(|p| p + 7))
-        .min()
-        .unwrap_or(rest.len());
-    format!(" {}", rest[..end].trim())
-}
-
-pub(crate) fn resolve_expr_type(
-    expr: &datafusion_expr::Expr,
-    input_schema: &Schema,
-    fallback_type: &DataType,
-) -> DataType {
-    match expr {
-        datafusion_expr::Expr::Column(col) => input_schema
-            .field_with_name(&col.name)
-            .map_or_else(|_| fallback_type.clone(), |f| f.data_type().clone()),
-        datafusion_expr::Expr::Literal(sv, _) => sv.data_type(),
-        datafusion_expr::Expr::Cast(cast) => cast.data_type.clone(),
-        datafusion_expr::Expr::TryCast(cast) => cast.data_type.clone(),
-        datafusion_expr::Expr::BinaryExpr(bin) => {
-            // For arithmetic, the result type is typically the wider
-            // of the two operands. Resolve the left side as a
-            // reasonable approximation.
-            resolve_expr_type(&bin.left, input_schema, fallback_type)
-        }
-        datafusion_expr::Expr::ScalarFunction(func) => {
-            // Try to get return type from input types
-            let arg_types: Vec<DataType> = func
-                .args
-                .iter()
-                .map(|a| resolve_expr_type(a, input_schema, fallback_type))
-                .collect();
-            func.func
-                .return_type(&arg_types)
-                .unwrap_or_else(|_| fallback_type.clone())
-        }
-        #[allow(deprecated)]
-        datafusion_expr::Expr::Wildcard { .. } => DataType::Boolean,
-        _ => fallback_type.clone(),
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    #[test]
-    fn test_interval_day_time_seconds_only() {
-        use arrow::datatypes::IntervalDayTime;
-        let sv = ScalarValue::IntervalDayTime(Some(IntervalDayTime::new(0, 10_000)));
-        let sql = scalar_value_to_sql(&sv);
-        assert_eq!(sql, "INTERVAL '10 seconds'");
-    }
-
-    #[test]
-    fn test_interval_day_time_days_only() {
-        use arrow::datatypes::IntervalDayTime;
-        let sv = ScalarValue::IntervalDayTime(Some(IntervalDayTime::new(3, 0)));
-        let sql = scalar_value_to_sql(&sv);
-        assert_eq!(sql, "INTERVAL '3 days'");
-    }
-
-    #[test]
-    fn test_interval_day_time_mixed() {
-        use arrow::datatypes::IntervalDayTime;
-        let sv = ScalarValue::IntervalDayTime(Some(IntervalDayTime::new(1, 5_500)));
-        let sql = scalar_value_to_sql(&sv);
-        assert_eq!(sql, "INTERVAL '1 days 5.500 seconds'");
-    }
-
-    #[test]
-    fn test_interval_year_month() {
-        let sv = ScalarValue::IntervalYearMonth(Some(15));
-        let sql = scalar_value_to_sql(&sv);
-        assert_eq!(sql, "INTERVAL '1 years 3 months'");
-    }
-
-    #[test]
-    fn test_interval_month_day_nano() {
-        use arrow::datatypes::IntervalMonthDayNano;
-        let sv =
-            ScalarValue::IntervalMonthDayNano(Some(IntervalMonthDayNano::new(2, 1, 3_000_000_000)));
-        let sql = scalar_value_to_sql(&sv);
-        assert_eq!(sql, "INTERVAL '2 months 1 days 3 seconds'");
-    }
 
     #[tokio::test]
     async fn test_try_from_sql_rejects_post_aggregate_projection() {
